requirements.txt
python-telegram-bot==21.5
groq==0.11.0
flask==3.0.3
python-dotenv==1.0.1
httpx==0.27.0